# substationDataServer/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Public Substation API for the Smart Service Intelligent Autonomous Grid",
    "description" : "Public API for registered users to access the device data, production data on a device, bay or substation base.",
    "contact" : {
      "name" : "Istvan Mudrak",
      "url" : "https://www.smartservice.energy",
      "email" : "istvan.mudrak@smartservice.energy"
    },
    "license" : {
      "name" : "Proprietary",
      "url" : "https://www.smartservice.energy/licenses/Substation"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.smartservice.energy/substation/v1"
  } ],
...
  "paths" : {
...
    "/measurements" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getMeasurementTypes",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total list of different timeseries of a substation could produce for a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeasurementTypes"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/actual/all" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getProductionAll",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Id of the series reported by the getMeasurementTypes",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total actual production of substations owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/actual/substation/{substationId}" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getProductionBySubstationId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Id of the series reported by the getMeasurementTypes",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "substationId",
          "in" : "path",
          "description" : "Id of the substation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total actual production of a substation owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/actual/bay/{bayId}" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getProductionByBayId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Id of the series reported by the getMeasurementTypes",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "bayId",
          "in" : "path",
          "description" : "Id of a given bay",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total actual production of a given bay",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/actual/device/{deviceId}" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getProductionByDeviceId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Datacom Products, CR0M3GP4U, Electrical Transceiver",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "deviceId",
          "in" : "path",
          "description" : "Id of a given device",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total actual production of a given device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/historical/all" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getMeasurementAll",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Datacom Products, CR0M3GP4U, Electrical Transceiver",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "fromDate",
          "in" : "query",
          "description" : "Start date of the historical production data",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "toDate",
          "in" : "query",
          "description" : "End date of the historical production data, when not provided the end date is the current date",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total historical production of substations owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/historical/substation/{substationId}" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getMeasurementsBySubstationId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Id of the series reported by the getMeasurementTypes",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "substationId",
          "in" : "path",
          "description" : "Id of a given substation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "fromDate",
          "in" : "query",
          "description" : "Start date of the historical production data",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "toDate",
          "in" : "query",
          "description" : "End date of the historical production data, when not provided the end date is the current date",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total historical production of the substations owned by a client",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/measurements/{measurementId}/historical/bay/{bayId}" : {
      "get" : {
        "tags" : [ "measurements" ],
        "operationId" : "getMeasurementsByBayId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "measurementId",
          "in" : "path",
          "description" : "Id of the series reported by the getMeasurementTypes",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "bayId",
          "in" : "path",
          "description" : "Id of a given bay",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "fromDate",
          "in" : "query",
          "description" : "Start date of the historical production data",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "toDate",
          "in" : "query",
          "description" : "End date of the historical production data, when not provided the end date is the current date",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Total historical production of a given bay",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeSeries"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        }
      },
...
      "MeasurementTypes" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/MeasurementType"
        }
      },
...
      "TimeSeries" : {
        "type" : "object",
        "properties" : {
          "typeid" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimeSeriesElement"
            }
          }
        },
        "example" : {
          "data" : [ {
            "date" : "2000-01-23T04:56:07.000+00:00",
            "value" : 0.80082819046101150206595775671303272247314453125
          }, {
            "date" : "2000-01-23T04:56:07.000+00:00",
            "value" : 0.80082819046101150206595775671303272247314453125
          } ],
          "typeid" : "typeid"
        }
      }
    },
    "securitySchemes" : {
      "smartservice_auth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/authorize",
            "tokenUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/token"
          }
        }
      }
    }
  }
}